backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-analytics-backend"
version = "0.1.0"
description = "Portfolio analytics MVP backend"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "greenlet>=3.1.1,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "redis>=5.2.0,<6.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.21.0,<1.0.0",
  "fakeredis>=2.31.0,<3.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.2.0,<2.0.0",
  "ruff>=0.11.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/api/*.py" = ["B008"]
"app/models/*.py" = ["F821", "UP037"]
"tests/conftest.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
